separated some functions to utilities.cpp
